digital_twin_app/requirements.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy
pydantic
python-dotenv
